refactor(tests): Move analyzeTaskComplexity tests to dedicated file
    - Created tests/unit/task-manager/analyze-task-complexity.test.js.
    - Moved tests for analyzeTaskComplexity from tests/unit/task-manager.test.js to the new dedicated file.
    - Adjusted mock setup in the new file to resolve test failures.
