+++ tests/unit/task-manager.test.js
@@ -324,306 +324,6 @@ describe('Task Manager Module', () => {
 		jest.clearAllMocks();
 	});
 
-	describe('findNextTask function', () => {
-		test('should return the highest priority task with all dependencies satisfied', () => {
-			const tasks = [
-				{
-					id: 1,
-					title: 'Setup Project',
-					status: 'done',
-					dependencies: [],
-					priority: 'high'
-				},
-				{
-					id: 2,
-					title: 'Implement Core Features',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'high'
-				},
-				{
-					id: 3,
-					title: 'Create Documentation',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'medium'
-				},
-				{
-					id: 4,
-					title: 'Deploy Application',
-					status: 'pending',
-					dependencies: [2, 3],
-					priority: 'high'
-				}
-			];
-
-			const nextTask = findNextTask(tasks);
-
-			expect(nextTask).toBeDefined();
-			expect(nextTask.id).toBe(2);
-			expect(nextTask.title).toBe('Implement Core Features');
-		});
-
-		test('should prioritize by priority level when dependencies are equal', () => {
-			const tasks = [
-				{
-					id: 1,
-					title: 'Setup Project',
-					status: 'done',
-					dependencies: [],
-					priority: 'high'
-				},
-				{
-					id: 2,
-					title: 'Low Priority Task',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'low'
-				},
-				{
-					id: 3,
-					title: 'Medium Priority Task',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'medium'
-				},
-				{
-					id: 4,
-					title: 'High Priority Task',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'high'
-				}
-			];
-
-			const nextTask = findNextTask(tasks);
-
-			expect(nextTask.id).toBe(4);
-			expect(nextTask.priority).toBe('high');
-		});
-
-		test('should return null when all tasks are completed', () => {
-			const tasks = [
-				{
-					id: 1,
-					title: 'Setup Project',
-					status: 'done',
-					dependencies: [],
-					priority: 'high'
-				},
-				{
-					id: 2,
-					title: 'Implement Features',
-					status: 'done',
-					dependencies: [1],
-					priority: 'high'
-				}
-			];
-
-			const nextTask = findNextTask(tasks);
-
-			expect(nextTask).toBeNull();
-		});
-
-		test('should return null when all pending tasks have unsatisfied dependencies', () => {
-			const tasks = [
-				{
-					id: 1,
-					title: 'Setup Project',
-					status: 'pending',
-					dependencies: [2],
-					priority: 'high'
-				},
-				{
-					id: 2,
-					title: 'Implement Features',
-					status: 'pending',
-					dependencies: [1],
-					priority: 'high'
-				}
-			];
-
-			const nextTask = findNextTask(tasks);
-
-			expect(nextTask).toBeNull();
-		});
-
-		test('should handle empty tasks array', () => {
-			const nextTask = findNextTask([]);
-
-			expect(nextTask).toBeNull();
-		});
-	});
-
-	describe('analyzeTaskComplexity function', () => {
-		// Setup common test variables
-		const tasksPath = 'tasks/tasks.json';
-		const reportPath = 'scripts/task-complexity-report.json';
-		const thresholdScore = 5;
-		const baseOptions = {
-			file: tasksPath,
-			output: reportPath,
-			threshold: thresholdScore.toString(),
-			research: false // Default to false
-		};
-
-		// Sample response structure (simplified for these tests)
-		const sampleApiResponse = {
-			tasks: [
-				{ id: 1, complexity: 3, subtaskCount: 2 },
-				{ id: 2, complexity: 7, subtaskCount: 5 },
-				{ id: 3, complexity: 9, subtaskCount: 8 }
-			]
-		};
-
-		beforeEach(() => {
-			jest.clearAllMocks();
-
-			// Setup default mock implementations
-			mockReadJSON.mockReturnValue(JSON.parse(JSON.stringify(sampleTasks)));
-			mockWriteJSON.mockImplementation((path, data) => data); // Return data for chaining/assertions
-			// Just set the mock resolved values directly - no spies needed
-			mockCallClaude.mockResolvedValue(sampleApiResponse);
-			mockCallPerplexity.mockResolvedValue(sampleApiResponse);
-
-			// Mock console methods to prevent test output clutter
-			jest.spyOn(console, 'log').mockImplementation(() => {});
-			jest.spyOn(console, 'error').mockImplementation(() => {});
-		});
-
-		afterEach(() => {
-			// Restore console methods
-			console.log.mockRestore();
-			console.error.mockRestore();
-		});
-
-		test('should call Claude when research flag is false', async () => {
-			// Arrange
-			const options = { ...baseOptions, research: false };
-
-			// Act
-			await testAnalyzeTaskComplexity(options);
-
-			// Assert
-			expect(mockCallClaude).toHaveBeenCalled();
-			expect(mockCallPerplexity).not.toHaveBeenCalled();
-			expect(mockWriteJSON).toHaveBeenCalledWith(
-				reportPath,
-				expect.any(Object)
-			);
-		});
-
-		test('should call Perplexity when research flag is true', async () => {
-			// Arrange
-			const options = { ...baseOptions, research: true };
-
-			// Act
-			await testAnalyzeTaskComplexity(options);
-
-			// Assert
-			expect(mockCallPerplexity).toHaveBeenCalled();
-			expect(mockCallClaude).not.toHaveBeenCalled();
-			expect(mockWriteJSON).toHaveBeenCalledWith(
-				reportPath,
-				expect.any(Object)
-			);
-		});
-
-		test('should handle valid JSON response from LLM (Claude)', async () => {
-			// Arrange
-			const options = { ...baseOptions, research: false };
-
-			// Act
-			await testAnalyzeTaskComplexity(options);
-
-			// Assert
-			expect(mockReadJSON).toHaveBeenCalledWith(tasksPath);
-			expect(mockCallClaude).toHaveBeenCalled();
-			expect(mockCallPerplexity).not.toHaveBeenCalled();
-			expect(mockWriteJSON).toHaveBeenCalledWith(
-				reportPath,
-				expect.objectContaining({
-					complexityAnalysis: expect.arrayContaining([
-						expect.objectContaining({ taskId: 1 })
-					])
-				})
-			);
-			expect(mockLog).toHaveBeenCalledWith(
-				'info',
-				expect.stringContaining('Successfully analyzed')
-			);
-		});
-
-		test('should handle and fix malformed JSON string response (Claude)', async () => {
-			// Arrange
-			const malformedJsonResponse = {
-				tasks: [{ id: 1, complexity: 3 }]
-			};
-			mockCallClaude.mockResolvedValueOnce(malformedJsonResponse);
-			const options = { ...baseOptions, research: false };
-
-			// Act
-			await testAnalyzeTaskComplexity(options);
-
-			// Assert
-			expect(mockCallClaude).toHaveBeenCalled();
-			expect(mockCallPerplexity).not.toHaveBeenCalled();
-			expect(mockWriteJSON).toHaveBeenCalled();
-		});
-
-		test('should handle missing tasks in the response (Claude)', async () => {
-			// Arrange
-			const incompleteResponse = { tasks: [sampleApiResponse.tasks[0]] };
-			mockCallClaude.mockResolvedValueOnce(incompleteResponse);
-
-			const options = { ...baseOptions, research: false };
-
-			// Act
-			await testAnalyzeTaskComplexity(options);
-
-			// Assert
-			expect(mockCallClaude).toHaveBeenCalled();
-			expect(mockCallPerplexity).not.toHaveBeenCalled();
-			expect(mockWriteJSON).toHaveBeenCalled();
-		});
-
-		// Add a new test specifically for threshold handling
-		test('should handle different threshold parameter types correctly', async () => {
-			// Test with string threshold
-			let options = { ...baseOptions, threshold: '7' };
-			const report1 = await testAnalyzeTaskComplexity(options);
-			expect(report1.meta.thresholdScore).toBe(7);
-			expect(mockCallClaude).toHaveBeenCalled();
-
-			// Reset mocks
-			jest.clearAllMocks();
-
-			// Test with number threshold
-			options = { ...baseOptions, threshold: 8 };
-			const report2 = await testAnalyzeTaskComplexity(options);
-			expect(report2.meta.thresholdScore).toBe(8);
-			expect(mockCallClaude).toHaveBeenCalled();
-
-			// Reset mocks
-			jest.clearAllMocks();
-
-			// Test with float threshold
-			options = { ...baseOptions, threshold: 6.5 };
-			const report3 = await testAnalyzeTaskComplexity(options);
-			expect(report3.meta.thresholdScore).toBe(6.5);
-			expect(mockCallClaude).toHaveBeenCalled();
-
-			// Reset mocks
-			jest.clearAllMocks();
-
-			// Test with undefined threshold (should use default)
-			const { threshold, ...optionsWithoutThreshold } = baseOptions;
-			const report4 = await testAnalyzeTaskComplexity(optionsWithoutThreshold);
-			expect(report4.meta.thresholdScore).toBe(5); // Default value from the function
-			expect(mockCallClaude).toHaveBeenCalled();
-		});
-	});
-
 	describe('parsePRD function', () => {
 		// Mock the sample PRD content
 		const samplePRDContent = '# Sample PRD for Testing';
@@ -813,7 +513,7 @@ describe('Task Manager Module', () => {
 		});
 	});
 
-	describe.skip('updateTasks function', () => {
+	describe('updateTasks function', () => {
 		test('should update tasks based on new context', async () => {
 			// This test would verify that:
 			// 1. The function reads the tasks file correctly
@@ -3101,68 +2801,3 @@ describe.skip('updateSubtaskById function', () => {
 
 	// More tests will go here...
 });
-
-// Add this test-specific implementation after the other test functions like testParsePRD
-const testAnalyzeTaskComplexity = async (options) => {
-	try {
-		// Get base options or use defaults
-		const thresholdScore = parseFloat(options.threshold || '5');
-		const useResearch = options.research === true;
-		const tasksPath = options.file || 'tasks/tasks.json';
-		const reportPath = options.output || 'scripts/task-complexity-report.json';
-		const modelName = options.model || 'mock-claude-model';
-
-		// Read tasks file
-		const tasksData = mockReadJSON(tasksPath);
-		if (!tasksData || !Array.isArray(tasksData.tasks)) {
-			throw new Error(`No valid tasks found in ${tasksPath}`);
-		}
-
-		// Filter tasks for analysis (non-completed)
-		const activeTasks = tasksData.tasks.filter(
-			(task) => task.status !== 'done' && task.status !== 'completed'
-		);
-
-		// Call the appropriate mock API based on research flag
-		let apiResponse;
-		if (useResearch) {
-			apiResponse = await mockCallPerplexity();
-		} else {
-			apiResponse = await mockCallClaude();
-		}
-
-		// Format report with threshold check
-		const report = {
-			meta: {
-				generatedAt: new Date().toISOString(),
-				tasksAnalyzed: activeTasks.length,
-				thresholdScore: thresholdScore,
-				projectName: tasksData.meta?.projectName || 'Test Project',
-				usedResearch: useResearch,
-				model: modelName
-			},
-			complexityAnalysis:
-				apiResponse.tasks?.map((task) => ({
-					taskId: task.id,
-					complexityScore: task.complexity || 5,
-					recommendedSubtasks: task.subtaskCount || 3,
-					expansionPrompt: `Generate ${task.subtaskCount || 3} subtasks`,
-					reasoning: 'Mock reasoning for testing'
-				})) || []
-		};
-
-		// Write the report
-		mockWriteJSON(reportPath, report);
-
-		// Log success
-		mockLog(
-			'info',
-			`Successfully analyzed ${activeTasks.length} tasks with threshold ${thresholdScore}`
-		);
-
-		return report;
-	} catch (error) {
-		mockLog('error', `Error during complexity analysis: ${error.message}`);
-		throw error;
-	}
-};
